--- AwesomeTSProject/containers/AddCrypto/styles.ts
@@ -1,42 +1,0 @@
-import styled from 'styled-components/native';
-
-export const CryptoInput = styled.TextInput`
-  border: 2px solid #fbd24d;
-  width: 90%;
-  height: 8%;
-  position: relative;
-  top: 38%;
-  margin: 5%;
-`;
-
-export const AddText = styled.Text`
-  fontSize: 24px;
-  color: black;
-  position: relative;
-  top: 35%;
-  left: 6%;
-  fontWeight: bold;
-`;
-
-export const backtext = styled.Text`
-    fontSize: 20px,
-    lineHeight: 21px,
-    fontWeight: bold,
-    letterSpacing: 0.25,
-    color: gray,
-`;
-
-export const modaltext = styled.Text`
-    fontSize: 16px,
-    lineHeight: 21px,
-    fontWeight: bold,
-    letterSpacing: 0.25,
-    color: gray,
-`;
-export const addBtext = styled.Text`
-    fontSize: 16px,
-    lineHeight: 21px,
-    fontWeight: bold,
-    letterSpacing: 0.25,
-    color: gray,
-`;
