+++ AwesomeTSProject/src/components/SingleCrypto/Styles.ts
@@ -1,4 +1,5 @@
 import styled from 'styled-components/native';
+import {theme} from '../../utils/theme';
 
 export const OneCrypto = styled.View`
   width: 100%;
@@ -14,14 +15,14 @@ export const NamesContainer = styled.View`
 `;
 
 export const CryptoName = styled.Text`
-  color: black;
+  color: ${theme.colors.black};
   fontSize: 20px;
   position:relative;
   bottom:35px;
 `;
 
 export const CryptoSymbol = styled.Text`
-  color: gray;
+  color: ${theme.colors.gray};
   fontSize: 16px;
   position:relative;
   bottom:35px;
@@ -38,7 +39,7 @@ export const CryptoLogo = styled.Image`
 export const ActualValue = styled.Text`
   position: relative;
   fontSize: 20px;
-  color: black;
+  color: ${theme.colors.black};
   left: 75%;
   bottom: 65%;
 `;
@@ -46,7 +47,7 @@ export const ActualValue = styled.Text`
 export const Increase = styled.Text`
   position: relative;
   fontSize: 14px;
-  color: green;
+  color: ${theme.colors.green};
   left: 82%;
   bottom: 55%;
 `;
@@ -54,17 +55,9 @@ export const Increase = styled.Text`
 export const Decrease = styled.Text`
   position: relative;
   fontSize: 14px;
-  color: red;
+  color: ${theme.colors.red};
   left: 82%;
   bottom: 55%;
-`;
-
-export const IncArrow = styled.Image`
-
-`;
-
-export const DecArrow = styled.Image`
- 
 `;
 
 export const ArrowCont = styled.View`
