+++ AwesomeTSProject/src/containers/CryptoList/styles.ts
@@ -1,9 +1,5 @@
 import styled from 'styled-components/native';
 
 export const Scrolls = styled.ScrollView`
-    Height:100%,
-`;
-
-export const ViewCont = styled.View`
-    Flex:1,
+    Height:90%,
 `;
