+++ AwesomeTSProject/src/react-app-env.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="react-scripts" />;
+
+declare module '*.png' {
+  export default '' as any;
+}
+declare module '*.svg' {
+  export default '' as any;
+}
+declare module '*.jpeg' {
+  export default '' as any;
+}
+declare module '*.jpg' {
+  export default '' as any;
+}
