+++ AwesomeTSProject/src/utils/theme.ts
@@ -1,0 +1,11 @@
+export const theme = {
+  colors: {
+    black: 'black',
+    white: '#fff',
+    gray: 'gray',
+    green: 'green',
+    red: 'red',
+    TopBarColor: '#385775',
+    CryptoInputColor: '#fbd24d',
+  },
+};
