+++ AwesomeTSProject/store/index.ts
@@ -1,5 +1,4 @@
 import {configureStore} from '@reduxjs/toolkit';
-// import RootReducer from './reducer/RootReducer';
 import cryptoSlice from './actions/cryptoSlice';
 
 export const store = configureStore({
@@ -8,7 +7,6 @@ export const store = configureStore({
   },
 });
 
-// Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>;
-// Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
+
 export type AppDispatch = typeof store.dispatch;
